+++ addon-test-support/dom-helpers.js
@@ -1,0 +1,1 @@
+export * from 'ember-cli-yadda-opinionated/test-support/-private/dom-helpers';