+++ addon-test-support/index.js
@@ -5,34 +5,6 @@ import {
   REGEX_LABEL
 } from 'ember-cli-yadda-opinionated/test-support/-private/regex';
 
-import {
-  clickByLabel,
-  doubleClickByLabel,
-  fillInByLabel,
-  findAllByLabel,
-  findByLabel,
-  findSingleByLabel,
-  findInputForLabelWithText,
-  mouseEnterByLabel,
-  mouseLeaveByLabel,
-  powerSelectFindTrigger,
-  powerSelectFindDropdown,
-  powerSelectFindOptions,
-  powerSelectFindSelectedOptions,
-  powerSelectFilterSelectedOptionsByText,
-  powerSelectIsSelectedOptionDisabled,
-  powerSelectRemoveSelectedOption,
-  powerSelectIsDropdownExpanded,
-  powerselectIsTriggerDisabled,
-  powerSelectExpand,
-  powerSelectCollapse,
-  powerSelectFindOptionByValueOrSelector,
-  powerDatePickerFindTrigger,
-  powerDatePickerFindDropdown,
-  triggerByLabel,
-  triggerKeyByLabel
-} from 'ember-cli-yadda-opinionated/test-support/-private/dom-helpers';
-
 import { labelMap, transformsMap } from 'ember-cli-yadda-opinionated/test-support/-private/maps';
 import selectorFromLabel from 'ember-cli-yadda-opinionated/test-support/-private/selector-from-label';
 import setupDictionary from 'ember-cli-yadda-opinionated/test-support/-private/setup-dictionary';
@@ -43,40 +15,15 @@ import whenSteps from 'ember-cli-yadda-opinionated/test-support/-private/steps/w
 const opinionatedSteps = {...givenSteps, ...whenSteps, ...thenSteps};
 
 export {
-  clickByLabel,
-  doubleClickByLabel,
   composeSteps,
   element,
-  fillInByLabel,
-  findAllByLabel,
-  findByLabel,
-  findInputForLabelWithText,
-  findSingleByLabel,
   givenSteps,
   labelMap,
-  mouseEnterByLabel,
-  mouseLeaveByLabel,
   opinionatedSteps,
-  powerDatePickerFindDropdown,
-  powerSelectFindTrigger,
-  powerSelectFindDropdown,
-  powerSelectFindOptions,
-  powerSelectFindSelectedOptions,
-  powerSelectFilterSelectedOptionsByText,
-  powerSelectIsSelectedOptionDisabled,
-  powerSelectRemoveSelectedOption,
-  powerSelectIsDropdownExpanded,
-  powerselectIsTriggerDisabled,
-  powerSelectExpand,
-  powerSelectCollapse,
-  powerSelectFindOptionByValueOrSelector,
-  powerDatePickerFindTrigger,
   REGEX_LABEL,
   selectorFromLabel,
   setupDictionary,
   thenSteps,
   transformsMap,
-  triggerByLabel,
-  triggerKeyByLabel,
   whenSteps
 };